chore(workspace,clippy): 6 more autofix wins across 6 crates
Same lint families as previous chore batches, swept across more files:
- bottleneck-exact/hopcroft_karp.rs:44 — collapsible_if: nested
  'if cond { if cond2 { ... } }' to combined 'if cond && cond2 { ... }'
- cli/bin/evaporchain-rehearse.rs:138 — needless format-arg for a
  string literal in println! — inline the unicode char directly
- consensus/tendermint.rs:2345 — manual_unwrap_or_default
  (re-fix; parallel session reverted line earlier)
- consensus/tendermint.rs:5955 — manual_is_multiple_of (re-fix)
- cslc/cssr.rs:637 — '1 <= df && df <= 5' to '(1..=5).contains(&df)'
  (clippy::manual_range_contains)
- happy-tensor/disk.rs:57 — '.or_insert_with(BTreeSet::new)' to
  '.or_default()' (clippy::unwrap_or_default)
- singh-counsel/intent.rs — same lint family

780 tests across 5 lib targets still green.

+++ crates/evaporchain-singh-counsel/src/intent.rs
@@ -145,7 +145,7 @@ impl Intent {
         h.update(b"evaporchain:counsel-intent:v1\0");
         h.update(&self.id.0);
         h.update(&self.holder);
-        h.update(&[self.verb as u8 as u8]);
+        h.update(&[(self.verb as u8)]);
         h.update(&(self.object.len() as u32).to_le_bytes());
         h.update(&self.object);
         h.update(&(self.constraint.len() as u32).to_le_bytes());
